refactor(validation): move validateRequest middleware to a new file and update import paths
